# openapi/v1.yaml
openapi: 3.1.0
info:
  title: Coziyoo API v1
  version: 1.0.0
  description: Versioned API contract for Coziyoo marketplace backend.
servers:
  - url: http://localhost:3000
    description: Local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /v1/seller/lots/{lotId}/adjust:
    post:
      tags: [lots]
      parameters:
        - in: path
          name: lotId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Lot adjusted }
  /v1/seller/lots/{lotId}/recall:
    post:
      tags: [lots]
      parameters:
        - in: path
          name: lotId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Lot recalled }
...
  /v1/admin/lots/{lotId}/orders:
    get:
      tags: [lots]
      parameters:
        - in: path
          name: lotId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200': { description: Orders affected by lot }
...